+++ src/entities/Sensor.hpp
@@ -19,7 +19,7 @@ using namespace std;
 class Sensor : public Entity {
   public:
     Sensor();
-    Sensor(int id, string name, Point* pos = NULL);
+    Sensor(long id, string name, Point* pos = NULL);
     Sensor(string name, Point* pos = NULL);
   
     Point*      getPosition();
